% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expit_f.R
\name{expit_f}
\alias{expit_f}
\title{Title
compute the expit}
\usage{
expit_f(dat, col)
}
\arguments{
\item{dat}{a data.frame}

\item{col}{the column that you want to compute expit}
}
\value{
the probability
}
\description{
Title
compute the expit
}
\examples{
mydata = data.frame(x = c(1:5))
expit_f(dat = mydata,x)
}
