% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toydata.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{Example data}
\format{
A data frame with 1028 rows and 4 variables:
\describe{
  \item{x1}{x1}
  \item{x2}{x2}
  \item{y}{y}
}
}
\source{
https://content.sph.harvard.edu/fitzmaur/ala2e/
}
\usage{
toydata
}
\description{
Data from a toxicology study blah blah blah
}
\references{
Price, C.J., Kimmel, C.A., Tyl, R.W. and Marr, M.C. (1985).
The developmental toxicity of ethylene glycol in rats and mice.
Toxicological Applications in Pharmacology, 81, 113-127.
}
\keyword{datasets}
